refactor(web): give MCP hooks one home
They were in three: lib/agents/hooks.ts, src/hooks/useMcpServers.ts, and
lib/tools/hooks.ts, whose only export was also an MCP hook.

Two of them were both called useMcpServers and returned the same shape from
different endpoints - /api/mcp/servers and /api/admin/mcp/servers. Importing
the wrong one type-checked and quietly answered a different question. They are
now useMcpServers and useAdminMcpServers, and every caller keeps the endpoint
it had before.

lib/tools is the right home rather than a new lib/mcp: MCP is one kind of tool
beside OpenAPI, and the domain is shared - ToolSnapshot spans both across 22
files, and the OAuth types are used by the OpenAPI modal.
