+++ web/src/lib/tools/hooks.ts
@@ -1,13 +1,102 @@
 "use client";
 
 import useSWR, { mutate } from "swr";
+import { useMemo } from "react";
 import { SWR_KEYS } from "@/lib/swr-keys";
 import { errorHandlingFetcher } from "@/lib/fetcher";
-import type { MCPServersResponse } from "@/lib/tools/interfaces";
+import type {
+  AgentEditorMCPServer,
+  MCPServersResponse,
+} from "@/lib/tools/interfaces";
 
-/** MCP servers an admin made available to Craft, with this user's connection
- * state (`craft_connected`). Distinct from `useMcpServers`, which reads the
- * admin-wide listing. */
+/**
+ * Every MCP server the current user can reach.
+ *
+ * This is the user-facing listing. For the admin console's view of every
+ * configured server, including ones this user cannot use, see
+ * {@link useAdminMcpServers} — the two return the same shape from different
+ * endpoints, so picking the wrong one type-checks and silently answers a
+ * different question.
+ */
+export function useMcpServers() {
+  const {
+    data: mcpData,
+    error,
+    isLoading,
+    mutate: mutateMcpServers,
+  } = useSWR<MCPServersResponse>(SWR_KEYS.mcpServers, errorHandlingFetcher);
+
+  return {
+    mcpData: mcpData ?? null,
+    isLoading,
+    error,
+    mutateMcpServers,
+  };
+}
+
+/**
+ * Every configured MCP server, from the admin endpoint. Use this only on admin
+ * surfaces; {@link useMcpServers} is what user-facing UI should read.
+ */
+export function useAdminMcpServers() {
+  const {
+    data: mcpData,
+    error,
+    isLoading,
+    mutate: mutateMcpServers,
+  } = useSWR<MCPServersResponse>(
+    SWR_KEYS.adminMcpServers,
+    errorHandlingFetcher
+  );
+
+  return {
+    mcpData: mcpData ?? null,
+    isLoading,
+    error,
+    mutateMcpServers,
+  };
+}
+
+/**
+ * The MCP servers relevant to one agent: those the user can reach, plus any
+ * already attached to the agent that they cannot. `can_attach` distinguishes
+ * them, so the editor can show an attached server without offering it as a
+ * choice the user is not allowed to make.
+ */
+export function useMcpServersForAgent(agentId: number | undefined) {
+  const accessible = useMcpServers();
+  const {
+    data: attachedData,
+    error: attachedError,
+    isLoading: attachedIsLoading,
+  } = useSWR<MCPServersResponse>(
+    agentId ? SWR_KEYS.agentMcpServers(agentId) : null,
+    errorHandlingFetcher
+  );
+
+  const mcpServers = useMemo<AgentEditorMCPServer[]>(() => {
+    const accessibleServers = accessible.mcpData?.mcp_servers ?? [];
+    const accessibleIds = new Set(accessibleServers.map((server) => server.id));
+    return [
+      ...accessibleServers.map((server) => ({ ...server, can_attach: true })),
+      ...(attachedData?.mcp_servers ?? [])
+        .filter((server) => !accessibleIds.has(server.id))
+        .map((server) => ({ ...server, can_attach: false })),
+    ];
+  }, [accessible.mcpData, attachedData]);
+
+  return {
+    mcpServers,
+    isLoading:
+      accessible.isLoading || (agentId !== undefined && attachedIsLoading),
+    error: accessible.error || attachedError,
+  };
+}
+
+/**
+ * MCP servers an admin made available to Craft, with this user's connection
+ * state (`craft_connected`).
+ */
 export function useCraftMcpServers(enabled: boolean = true) {
   const { data, error, isLoading } = useSWR<MCPServersResponse>(
     enabled ? SWR_KEYS.mcpServersCraft : null,
