+++ web/src/views/AgentEditorPage.tsx
@@ -70,11 +70,11 @@ import CustomAgentAvatar, {
 } from "@/refresh-components/avatars/CustomAgentAvatar";
 import InputAvatar from "@/refresh-components/inputs/InputAvatar";
 import SquareButton from "@/refresh-components/buttons/SquareButton";
-import { useAgents, useLabels } from "@/lib/agents/hooks";
+import { useAgents, useAgentLabels } from "@/lib/agents/hooks";
 import { createAgent, updateAgent } from "@/lib/agents/svc";
 import InputChipField from "@/refresh-components/inputs/InputChipField";
 import { AgentUpsertParameters } from "@/lib/agents/types";
-import { useMcpServersForAgent } from "@/lib/agents/hooks";
+import { useMcpServersForAgent } from "@/lib/tools/hooks";
 import useOpenApiTools from "@/hooks/useOpenApiTools";
 import { useAvailableTools } from "@/hooks/useAvailableTools";
 import { getActionIcon } from "@/lib/tools/mcpUtils";
@@ -573,7 +573,7 @@ export default function AgentEditorPage({
   const userFileIdsRef = useRef<string[]>([]);
 
   // Labels are edited in the Share Agent section and saved with the form
-  const { labels: allLabels, createLabel } = useLabels();
+  const { labels: allLabels, createLabel } = useAgentLabels();
   const [labelInputValue, setLabelInputValue] = useState("");
   const addAgentLabel = useCallback(
     async (
